third_party/WebKit/LayoutTests/html/dialog/inert-node-is-unselectable-expected.txt

Test that inert nodes cannot be selected. The test passes if the only text you can select is inside the dialog.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getSelection().toString() is "I'm selectable."
PASS successfullyParsed is true

TEST COMPLETE
Here is a text node you can't select. I'm selectable.
